llm-backend/requirements.txt
fastapi
uvicorn
python-dotenv

# Pyroscope profiling
pyroscope-io

# OpenTelemetry core + OTLP exporter
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0

# Instrumentations
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-openai

# OpenAI + LLM tracing
openai
traceloop-sdk>=0.6.0



